Separate common data and tarball backend
- Create separate file for common data backend
  functions
- Create separate file for tarball backend functions
- Put backward compatibility to common data backend
- Create a log directory for each data backup
